radiomicsfeatureextractionpipeline/backend/requirements.txt
aioredis==1.3.0
alabaster==0.7.10
altgraph==0.16.1
appdirs==1.4.3
asgiref==3.2.2
asn1crypto==0.24.0
astroid==2.2.5
astropy==3.0.2
async-timeout==3.0.1
attrs==18.1.0
autobahn==19.10.1
Automat==0.7.0
autopep8==1.4.4
Babel==2.5.3
backcall==0.1.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
cffi==1.12.3
channels==2.3.0
channels-redis==2.4.0
colorama==0.4.1
constantly==15.1.0
cryptography==2.7
daphne==2.3.0
decorator==4.4.0
dill==0.3.1.1
Django==2.2.5
django-cors-headers==3.1.0
django-filter==2.2.0
djangorestframework==3.10.3
docopt==0.6.2
h2==3.1.1
hiredis==1.0.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==19.0.0
idna==2.8
imageio==2.5.0
incremental==17.5.0
lazy-object-proxy==1.3.1
msgpack==0.6.2
multiprocess==0.70.9
networkx==2.3
numpy==1.17.2
pandas==0.25.1
pathos==0.2.5
Pillow==6.1.0
pox==0.2.7
ppft==1.6.6.1
priority==1.3.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycodestyle==2.4.0
pycparser==2.19
pydicom==1.3.0
pygubu==0.9.8.2
PyHamcrest==1.9.0
pykwalify==1.7.0
pyOpenSSL==19.0.0
pyradiomics==2.1.2
python-dateutil==2.8.0
pytz==2018.4
PyWavelets==1.0.0
PyYAML==5.1.2
scikit-image==0.15.0
scipy==1.3.1
service-identity==18.1.0
SimpleITK==1.2.2
six==1.11.0
sqlparse==0.3.0
Twisted==19.7.0
txaio==18.8.1
typed-ast==1.4.0
wrapt==1.10.11
zope.interface==4.6.0
